lawrencetrailhawks/requirements/base.txt
DateUtils
Django>=1.7,<1.8
Markdown==2.6.5
Unipath==1.1
akismet==0.2.0
boto==2.36.0
click==3.3
django-adaptors==0.2.5
django-ajaximage==0.2.7
django-backupdb==0.6.0
django-braces==1.4.0
django-contrib-comments==1.5
django-docopt-command
django-extensions
django-favicon==0.1.2
django-filter==0.9.2
django-gravatar2==1.1.4
django-hosts==1.1
django-ical==1.3
django-markup-deprecated==0.0.3
django-registration==1.0
django-robots==1.0
django-shorturls==1.1.0
django-simple-open-graph==0.3
django-storages==1.1.8
django-taggit==0.18.0
django-thumbor==0.5
django-waffle==0.10.1
djangorestframework==3.1.1
djrill==0.7.0
feedparser==5.1.3
flickrapi==2.0
gunicorn==19.1.1
httplib2==0.9
lxml
micawber==0.3.3
num2words==0.5.3
psycopg2==2.5.4
raven==5.1.1
recaptcha-client==1.0.6
requests<3.0.0
simplejson==3.6.5
titlecase==0.8.0

#git+https://bitbucket.org/webology/django-shorturls.git#egg=django-shorturls-git
git+https://github.com/jefftriplett/django-syncr.git#egg=django-syncr-git
git+https://github.com/vikingco/django-storage-migration.git#egg=django-storage-migration-git
